[#2] Move citations fn to utils & improve citations style
